chore: move handlers into helpers
