+++ helpers/github.ts
@@ -3,9 +3,19 @@ import { Octokit } from "@octokit/rest";
 import _projects from "../projects.json";
 import opt from "../opt.json";
 import { Statistics } from "../types/statistics";
+import { writeFile } from "fs/promises";
+import twitter from "./twitter";
 
 export type GitHubIssue = RestEndpointMethodTypes["issues"]["get"]["response"]["data"];
 export type GitHubLabel = RestEndpointMethodTypes["issues"]["listLabelsOnIssue"]["response"]["data"][0];
+
+export type StateChanges<T extends string = "open" | "closed"> = {
+  [key: string]: {
+    cause: boolean;
+    effect: T;
+    comment: string;
+  };
+};
 
 export const projects = _projects as {
   urls: string[];
@@ -360,5 +370,158 @@ export async function writeTotalRewardsToGithub(statistics: Statistics) {
   } catch (error) {
     console.error(`Error writing total rewards to github file: ${error}`);
     throw error;
+  }
+}
+
+export async function createDevPoolIssue(projectIssue: GitHubIssue, projectUrl: string, body: string, twitterMap: { [key: string]: string }) {
+  // if issue is "closed" then skip it, no need to copy/paste already "closed" issues
+  if (projectIssue.state == "closed") return;
+
+  // if the project issue is assigned to someone, then skip it
+  if (projectIssue.assignee?.login) return;
+
+  // if issue doesn't have the "Price" label then skip it, no need to pollute repo with draft issues
+  if (!(projectIssue.labels as GitHubLabel[]).some((label) => label.name.includes(LABELS.PRICE))) return;
+
+  // create a new issue
+  const createdIssue = await octokit.rest.issues.create({
+    owner: DEVPOOL_OWNER_NAME,
+    repo: DEVPOOL_REPO_NAME,
+    title: projectIssue.title,
+    body,
+    labels: getDevpoolIssueLabels(projectIssue, projectUrl),
+  });
+  console.log(`Created: ${createdIssue.data.html_url} (${projectIssue.html_url})`);
+
+  // post to social media
+  const socialMediaText = getSocialMediaText(createdIssue.data);
+  const tweetId = await twitter.postTweet(socialMediaText);
+
+  twitterMap[createdIssue.data.node_id] = tweetId?.id ?? "";
+  await writeFile("./twitterMap.json", JSON.stringify(twitterMap));
+}
+
+export async function handleDevPoolIssue(
+  projectIssues: GitHubIssue[],
+  projectIssue: GitHubIssue,
+  projectUrl: string,
+  devpoolIssue: GitHubIssue,
+  body: string,
+  isFork: boolean
+) {
+  const metaChanges = {
+    // the title of the issue has changed
+    title: devpoolIssue.title != projectIssue.title,
+    // the issue url has updated
+    body: devpoolIssue.body != projectIssue.html_url,
+    // the price/priority labels have changed
+    labels:
+      (devpoolIssue.labels as GitHubLabel[])
+        .map((label) => label.name)
+        .sort()
+        .toString() != getDevpoolIssueLabels(projectIssue, projectUrl).sort().toString(),
+  };
+
+  // process only the metadata changes
+  // forked body will always be different because of the www
+  if (metaChanges.title || (!isFork && metaChanges.body) || metaChanges.labels) {
+    try {
+      await octokit.rest.issues.update({
+        owner: DEVPOOL_OWNER_NAME,
+        repo: DEVPOOL_REPO_NAME,
+        issue_number: devpoolIssue.number,
+        title: projectIssue.title,
+        body,
+        labels: getDevpoolIssueLabels(projectIssue, projectUrl),
+      });
+    } catch (err) {
+      console.error(err);
+    }
+
+    console.log(
+      `Updated [${(metaChanges.title && '"title", ', metaChanges.labels && '"labels", ', metaChanges.body && '"body"]')}:\nDevpool Issue: ${
+        devpoolIssue.html_url
+      }\nProject Issue: ${projectIssue.html_url}`
+    );
+  }
+
+  const hasNoPriceLabels = !(projectIssue.labels as GitHubLabel[]).some((label) => label.name.includes(LABELS.PRICE));
+
+  // these changes will open/close issues
+  const stateChanges: StateChanges = {
+    // missing in the partners
+    forceMissing_Close: {
+      cause: !projectIssues.some((projectIssue) => projectIssue.node_id == getIssueLabelValue(devpoolIssue, "id:")),
+      effect: "closed",
+      comment: "Closed (missing in partners):",
+    },
+    // no price labels set and open in the devpool
+    noPriceLabels_Close: {
+      cause: hasNoPriceLabels && devpoolIssue.state == "open",
+      effect: "closed",
+      comment: "Closed (no price labels):",
+    },
+    // it's closed, been merged and still open in the devpool
+    issueComplete_Close: {
+      cause: projectIssue.state == "closed" && devpoolIssue.state == "open" && !!projectIssue.pull_request?.merged_at,
+      effect: "closed",
+      comment: "Closed (merged):",
+    },
+    // it's closed, not merged and still open in the devpool
+    issueClosed_Close: {
+      cause: projectIssue.state == "closed" && devpoolIssue.state == "open",
+      effect: "closed",
+      comment: "Closed (not merged):",
+    },
+    // it's closed, assigned and still open in the devpool
+    issueAssignedClosed_Close: {
+      cause: projectIssue.state == "closed" && devpoolIssue.state == "open" && !!projectIssue.assignee?.login,
+      effect: "closed",
+      comment: "Closed (assigned-closed):",
+    },
+    // it's open, assigned and still open in the devpool
+    issueAssignedOpen_Close: {
+      cause: projectIssue.state == "open" && devpoolIssue.state == "open" && !!projectIssue.assignee?.login,
+      effect: "closed",
+      comment: "Closed (assigned-open):",
+    },
+    // it's open, unassigned and closed in the devpool
+    issueUnassigned_Open: {
+      cause: projectIssue.state == "open" && devpoolIssue.state == "closed" && !projectIssue.assignee?.login && !hasNoPriceLabels,
+      effect: "open",
+      comment: "Reopened (unassigned):",
+    },
+    // it's open, merged and closed in the devpool
+    issueReopenedMerged_Open: {
+      cause: projectIssue.state == "open" && devpoolIssue.state == "closed" && !!projectIssue.pull_request?.merged_at && !hasNoPriceLabels,
+      effect: "open",
+      comment: "Reopened (merged):",
+    },
+  };
+
+  let newState: "open" | "closed" | undefined = undefined;
+
+  // then process the state changes
+  for (const [, value] of Object.entries(stateChanges)) {
+    // if the cause is true and the effect is different from the current state
+    if (value.cause && devpoolIssue.state != value.effect) {
+      // if the new state is already set, then skip it
+      if (newState && newState == value.effect) continue;
+
+      try {
+        await octokit.rest.issues.update({
+          owner: DEVPOOL_OWNER_NAME,
+          repo: DEVPOOL_REPO_NAME,
+          issue_number: devpoolIssue.number,
+          state: value.effect,
+        });
+
+        newState = value.effect;
+      } catch (err) {
+        console.log(err);
+      }
+
+      console.log(`${value.comment}:\nDevpool Issue: ${devpoolIssue.html_url}\nProject Issue: ${projectIssue.html_url}`);
+    }
   }
 }
